BPR-54: refactor to delete temp of page close and add new service
