+++ BPRBE/Services/IDependencyComponentService.cs
@@ -4,12 +4,7 @@ namespace BPRBE.Services;
 
 public interface IDependencyComponentService {
 
-    /// <summary>
-    /// Loads the codebase in a temp folder
-    /// </summary>
-    /// <param name="file">the source code as a zip file</param>
-    /// <returns>the path to the temp folder as a string</returns>
-    public string LoadCodebaseInTemp(ArchiveFile file);
+   
     /// <summary>
     /// Gets the project names inside the loaded source code saved in a temp folder
     /// </summary>
