+++ BPRBlazor/Pages/Index.razor
@@ -3,6 +3,8 @@
 @using BPRBE.Services
 @inject IHttpService HttpService
 @inject IDependencyComponentService DependencyComponentService
+@inject ICodebaseService CodebaseService
+@implements IDisposable
 
 <button class="btn btn-primary" @onclick="SendDataAsync">Send Data</button>
 <DownloadButton />
