+++ BPRBlazor/Program.cs
@@ -9,7 +9,7 @@ builder.Services.AddHttpClient();
 
 builder.Services.AddScoped<IHttpService, HttpService>();
 builder.Services.AddScoped<IDependencyComponentService, DependencyComponentService>();
-
+builder.Services.AddScoped<ICodebaseService, CodebaseService>();
 builder.Services.AddSingleton(new JsonSerializerOptions
 {
     PropertyNameCaseInsensitive = true,
